refactor: Move ride classes from 'c2' package to 'ride' package for better organization
