+++ raptor/src/main/java/org/opentripplanner/raptor/rangeraptor/multicriteria/ride/PatternRideC1.java
@@ -1,8 +1,6 @@
-package org.opentripplanner.raptor.rangeraptor.multicriteria.ride.c1;
+package org.opentripplanner.raptor.rangeraptor.multicriteria.ride;
 
 import org.opentripplanner.raptor.rangeraptor.multicriteria.arrivals.stop.McStopArrival;
-import org.opentripplanner.raptor.rangeraptor.multicriteria.ride.PatternRide;
-import org.opentripplanner.raptor.rangeraptor.multicriteria.ride.PatternRideFactory;
 import org.opentripplanner.raptor.spi.RaptorCostCalculator;
 import org.opentripplanner.raptor.spi.RaptorTripSchedule;
 import org.opentripplanner.raptor.util.paretoset.ParetoComparator;
