+++ raptor/src/test/java/org/opentripplanner/raptor/rangeraptor/multicriteria/ride/PatternRideC1Test.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.raptor.rangeraptor.multicriteria.ride.c1;
+package org.opentripplanner.raptor.rangeraptor.multicriteria.ride;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
